something's going wrong...
so I added the two scaffolding things over here, but that's not the
problem.

Getting the meaning of selected referent is seriously malfunctioning.
Sometimes it works fine,
sometimes it shows the wrong meaning,
sometimes no selection leads to the next slide and gets a (obviously
wrong, because there shouldn't be a) meaning,
and sometimes it won't even proceed to the next slide.

WHAT'S GOING ON??